+++ authlib/jose/__init__.py
@@ -1,3 +1,11 @@
+"""
+    authlib.jose
+    ~~~~~~~~~~~~
+
+    JOSE implementation in Authlib. Tracking the status of JOSE specs at
+    https://tools.ietf.org/wg/jose/
+"""
+
 from .rfc7515 import (
     JWS, JWSAlgorithm, JWSHeader, JWSObject,
 )