+++ authlib/oauth1/__init__.py
@@ -1,23 +1,21 @@
 # coding: utf-8
 
-from .rfc5849.wrapper import OAuth1Request
-from .rfc5849.auth_client import AuthClient
-from .rfc5849.signature import (
+from .rfc5849 import (
+    OAuth1Request,
+    AuthClient,
     SIGNATURE_HMAC_SHA1,
     SIGNATURE_RSA_SHA1,
     SIGNATURE_PLAINTEXT,
     SIGNATURE_TYPE_HEADER,
     SIGNATURE_TYPE_QUERY,
     SIGNATURE_TYPE_BODY,
-)
-from .rfc5849.models import (
     ClientMixin,
     TemporaryCredentialMixin,
     TokenCredentialMixin,
     TemporaryCredential,
+    AuthorizationServer,
+    ResourceProtector,
 )
-from .rfc5849.authorization_server import AuthorizationServer
-from .rfc5849.resource_protector import ResourceProtector
 
 __all__ = [
     'OAuth1Request',