+++ authlib/specs/rfc7636/__init__.py
@@ -1,13 +1,2 @@
-# -*- coding: utf-8 -*-
-"""
-    authlib.specs.rfc7636
-    ~~~~~~~~~~~~~~~~~~~~~
-
-    This module represents a direct implementation of
-    Proof Key for Code Exchange by OAuth Public Clients.
-
-    https://tools.ietf.org/html/rfc7636
-"""
-
 from .grant import AuthorizationCodeGrant
 from .challenge import CodeChallenge, create_s256_code_challenge